#67: refactored inference-namespace
