+++ src/app/trajectory/inferences/inferences.page.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit } from '@angular/core'
 import { ActivatedRoute, Router } from '@angular/router'
 import { Inference } from 'src/app/model/inference'
-import { AllInferences } from 'src/app/shared-services/inferences/definitions'
-import { InferenceService } from './inference.service'
+import { AllInferences } from 'src/app/shared-services/inferences/engine/definitions'
+import { InferenceService } from 'src/app/shared-services/inferences/inference.service'
 
 @Component({
   selector: 'app-inferences',
